Change example sentence finder based on nlp
 Class FindSentenceIncludeWord in module find_sentence_include_word had
one problem. It was it can't find sentence properly when word is
included in sentence as inflected form(something like present tense,
plural form.).
 So I made a new class 'FindSentenceIncludeWordNlp' and modified
main_cont to use it.
<Additioanl changes>
1.The class 'FindSentenceIncludeWordNlp' uses nltk library. So, I
made requirements.txt.
2.Before this commit, a folder name that contains files something like
new words or example sentences was 'contexts'. I changed it to
'contents' because I think this name is more proper.