Change string splitter to use nltk
I removed ParseByDot and replaced it to SplitBySentence.

There is a problem in parse_by_dot. It was that it can't split sentences
appropriately that contains i.e., e.g. etc. It was drvied because this
class split sentences based on only dot. So I changed it to use nltk
which is based on nlp.
<additional changes>
change .gitignore to ignore .venv in parent folder not in
example_sentence_adder.