Modify FindSentenceIncludeWordNlp to be able to find word with spaces
 Prior version of FindSentenceIncludeWordNlp finds sentences based on
full original word. So It couldn't find word with spaces(ex.give a
way to ~). It was a bug.
 I modify FindSentenceIncludeWordNlp like it. When word with spaces
inputed, split it based on space, find based on each splitted, and if
there is a sentence that contains all splitted, add that sentence to be
finded. Also, in case of there is '~', '...', 'A', 'B' that indicate
space which new word can be inserted, remove those in splitted_word
list.